third_party/WebKit/LayoutTests/external/wpt/cookies/samesite/form-post-blank-reload-expected.txt

This is a testharness.js-based test.
PASS Reloaded same-host top-level form POSTs are strictly same-site
PASS Reloaded subdomain top-level form POSTs are strictly same-site
FAIL Reloaded cross-site top-level form POSTs are not same-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.7678069773408558" but got (undefined) undefined
Harness: the test ran to completion.
